ai-service/pyproject.toml
[project]
name = "finsight-ai-service"
version = "0.1.0"
description = "FinSight AI - RAG, strategy engine, chat, market layer"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "anthropic>=0.18.0",
    "httpx>=0.26.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.uv]
dev-dependencies = []
